% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{all_patients}
\alias{all_patients}
\title{Metadata of 8852 TCGA patients}
\format{A data frame with 8852 rows and 11 variables}
\usage{
data(all_patients)
}
\description{
A dataset containing the attributes of 8852 TCGA patients . The variables are as follows:
}
\details{
\itemize{
  \item Patient.
  \item disease2.
  \item age. Age at diagnosis.
  \item agez. Standardized age.
  \item race2. Race, some are predicted.
  \item EA. Whether european ancestry.
  \item gender.
  \item study_size.
  \item femme_ratio.
  \item participant. TCGA patient uid
}
}
\keyword{datasets}

