% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plot_style_misc.R
\name{expand_outer}
\alias{expand_outer}
\title{Fill all combination of the two input columns}
\usage{
expand_outer(data, X = "disease2", Y = "Gene")
}
\arguments{
\item{data}{input data}

\item{X}{column 1 Defaults to "disease2"}

\item{Y}{column 2 Defaults to "Gene"}
}
\description{
Form the outer product of the input column X,Y, typically disease2, Gene
}

