% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plot_style_misc.R
\name{extractPatient}
\alias{extractPatient}
\title{Extract patient names from input data}
\usage{
extractPatient(data)
}
\arguments{
\item{data}{input data}
}
\description{
input data maybe a vector or data frame
}

