% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plot_style_misc.R
\name{labelDisease}
\alias{labelDisease}
\title{Translate disease abbreviation}
\usage{
labelDisease(p, axis = "x", ...)
}
\arguments{
\item{p}{ggplot2 object}
}
\description{
Plot the all_mRNA levels of patients with variants in a specific gene
}

