% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{list_goi}
\alias{list_goi}
\title{Candidate gene list}
\format{A data frame with 1685 rows and 11 variables
\itemize{
  \item Gene.
  \item disease2.
}}
\usage{
list_goi
}
\description{
A dataset containing the attributes of 1680 candidate genes. The variables are as follows:
}
\keyword{datasets}

