% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{nsSNP_muts}
\alias{nsSNP_muts}
\title{Info on all identified variants}
\format{A data frame with 8852 rows and 11 variables
\itemize{
  \item uid.
  \item Patient.
  \item agez. Age at diagnosis.
}}
\usage{
nsSNP_muts
}
\description{
A dataset containing the attributes of 8852 TCGA patients . The variables are as follows:
}
\keyword{datasets}

