% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plot_style_misc.R
\name{nudgeXY}
\alias{nudgeXY}
\title{manually adjust a few genes' plotting coordinate}
\usage{
nudgeXY(df, Genes, axis, nudge)
}
\arguments{
\item{df}{input plotting data}

\item{Genes}{input Gene list}

\item{axis}{which axis}

\item{nudge}{amount}
}
\description{
Assume the plotting dataframe contains plot_x, plot_y for coordiantes
}

