% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plotMutRNASeq.R
\name{plotMutRNASeq}
\alias{plotMutRNASeq}
\title{Plot mutation RNASeq levels}
\usage{
plotMutRNASeq(data, Gene = NA, transformed = T)
}
\arguments{
\item{Gene}{Which gene to plot Defaults to NA.}

\item{transformed}{Whether to plot variance-stablized counts Defaults to TRUE.}

\item{df}{Supplied mutation df or a list of patients  Defaults to TRUE.}
}
\description{
Plot the all_mRNA levels of patients with variants in a specific gene
}
\keyword{RNASeq}

