% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/variant_handle_misc.R
\name{showLoF}
\alias{showLoF}
\title{Select info from LoF df}
\usage{
showLoF(data, show = T)
}
\arguments{
\item{data}{LoF variant df}

\item{show}{Whether to view the df Defaults to TRUE.}
}
\description{
select relevant infor mation from LoF df object
}

