% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/variant_handle_misc.R
\name{showMut}
\alias{showMut}
\title{Select info from mutation df}
\usage{
showMut(data)
}
\arguments{
\item{data}{mut df}
}
\description{
select relevant infor mation from LoF df object
}

