% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/variant_handle_misc.R
\name{showSNP}
\alias{showSNP}
\title{Select info from SNP df}
\usage{
showSNP(data, show = T)
}
\arguments{
\item{data}{SNP variant df}

\item{show}{Whether to view the df Defaults to TRUE.}
}
\description{
select relevant infor mation from SNP df object
}

