% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plot_style_misc.R
\name{sigSymbol}
\alias{sigSymbol}
\title{Label significance level for pvalues}
\usage{
sigSymbol(pval)
}
\arguments{
\item{pval}{input data}
}
\description{
Take in pval vector, return character vector
}

