% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plot_style_misc.R
\name{unifyPlotDim}
\alias{unifyPlotDim}
\title{Unify ggplot2 dimensions}
\usage{
unifyPlotDim(pl, width = T, height = T)
}
\arguments{
\item{width}{unify width data Defaults to TRUE}

\item{height}{unify height Defaults to TRUE}
}
\description{
Make ggplot2 objects the same width and height
}

